Harmonize with Helen DLT Framework

+++ src/xdbutils/pipelines/data.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+import warnings
 from pyspark.sql.functions import col
 from xdbutils.pipelines.management import DLTPipelineManager
 
@@ -10,17 +12,17 @@ class DLTPipelineDataManager(DLTPipelineManager):
         ids,
     ):
         pipeline_id = self._get_id()
-        assert pipeline_id, f"Pipeline {self.source_system}-{self.entity} not found"
+        assert pipeline_id, f"Pipeline {self.source_system}-{self.source_class} not found"
 
         update_id = self._get_latest_update(
             pipeline_id=pipeline_id,
         )
-        assert update_id, f"Pipeline {self.source_system}-{self.entity}: latest update not found"
+        assert update_id, f"Pipeline {self.source_system}-{self.source_class}: latest update not found"
 
         save_continuous_workflow = self.continuous_workflow
         if self.continuous_workflow:
-            print(f"Stopping pipeline {self.source_system}-{self.entity} and setting pipeline mode to triggered instead of continuous")
             self._stop(pipeline_id=pipeline_id)
+            print(f"Stopping pipeline {self.source_system}-{self.source_class} and setting pipeline mode to triggered instead of continuous")
             self.continuous_workflow = False
             self.create_or_update()
 
@@ -28,28 +30,49 @@ class DLTPipelineDataManager(DLTPipelineManager):
             pipeline_id=pipeline_id,
             update_id=update_id,
         )
-        assert datasets, f"Pipeline {self.source_system}-{self.entity}: cannot find datasets"
-        bronze_tables = [d for d in list(datasets) if d.startswith("bronze_")]
-        non_bronze_tables = [
-            d
-            for d in list(datasets)
-            if not d.startswith("bronze_") and not d.startswith("view_")
-        ]
+        assert datasets, f"Pipeline {self.source_system}-{self.source_class}: cannot find datasets"
 
+        ids_string = ",".join([f"'{id}'" for id in ids])
+        refresh_tables = []
+        bronze_updated = False
+        silver_updated = False
+
+        bronze_tables = [d for d in list(datasets) if d.startswith("bronze_")]
         for bronze_table in bronze_tables:
-            ids_string = ",".join([f"'{id}'" for id in ids])
-            statement = f"delete from {self.catalog}.{self.source_system}.{bronze_table} where `{id_column}` in ({ids_string})"
-            print(statement)
-            self.spark.sql(statement)
+            df = self.spark.sql(f"select * from {self.catalog}.{self.source_system}.{bronze_table}")
+            if any(column == id_column for column in df.columns):
+                bronze_updated = True
+                statement = f"delete from {self.catalog}.{self.source_system}.{bronze_table} where `{id_column}` in ({ids_string})"
+                print(statement)
+                self.spark.sql(statement)
+
+        if bronze_updated:
+            refresh_tables = [d for d in list(datasets) if not d.startswith("bronze_") and not d.startswith("view_")]
+        else:
+            silver_tables = [d for d in list(datasets) if d.startswith("silver_") and not d.startswith("view_")]
+            for silver_table in silver_tables:
+                df = self.spark.sql(f"select * from {self.catalog}.{self.source_system}.{silver_table}")
+                if any(column == id_column for column in df.columns):
+                    silver_updated = True
+                    statement = f"delete from {self.catalog}.{self.source_system}.{silver_table} where `{id_column}` in ({ids_string})"
+                    print(statement)
+                    self.spark.sql(statement)
         
-        print(f"Running pipeline {self.source_system}-{self.entity} with full refresh of: {', '.join(non_bronze_tables)}")
-        self._refresh(
-            pipeline_id=pipeline_id,
-            full_refresh_selection=non_bronze_tables,
-        )
+        if silver_updated:
+            refresh_tables = [d for d in list(datasets) if not d.startswith("bronze_") and not d.startswith("silver_") and not d.startswith("view_")]
+
+        if not bronze_updated and not silver_updated:
+            print(f"Column {id_column} not found in bronze or silver tables.")
+
+        if len(refresh_tables) > 0:
+            print(f"Running pipeline {self.source_system}-{self.source_class} with full refresh of: {', '.join(refresh_tables)}")
+            self._refresh(
+                pipeline_id=pipeline_id,
+                full_refresh_selection=refresh_tables,
+            )
 
         if self.continuous_workflow != save_continuous_workflow:
-            print(f"Updating pipeline {self.source_system}-{self.entity}, setting pipeline mode back to continuous")
+            print(f"Updating pipeline {self.source_system}-{self.source_class}, setting pipeline mode back to continuous")
             self.continuous_workflow = save_continuous_workflow
             self.create_or_update()
 
@@ -86,7 +109,6 @@ class DLTPipelineDataManager(DLTPipelineManager):
                         select *
                         from {self.catalog}.{schema}.{table}
                         """)
-                        print(f"""Created view {silver_catalog}.{schema}.{table.replace("silver_", "")}""")
                     if parts[0] == "gold":
                         self.spark.sql(f"create schema if not exists {gold_catalog}.{schema}")
                         self.spark.sql(f"""
@@ -95,4 +117,59 @@ class DLTPipelineDataManager(DLTPipelineManager):
                         select *
                         from {self.catalog}.{schema}.{table}
                         """)
-                        print(f"""Created view {silver_catalog}.{schema}.{table.replace("gold_", "")}""")
+
+    def backup_bronze(self, data_type = None, environment = None,
+        ):
+        """ Backup bronze data """
+        if data_type:
+            warnings.warn("Parameter data_type is deprecated",
+                category=DeprecationWarning,
+                stacklevel=2)
+        if environment:
+            warnings.warn("Parameter environment is deprecated",
+                category=DeprecationWarning,
+                stacklevel=2)
+        input_table = f"{self.catalog}.{self.source_system}.bronze_{self.source_class}"
+        ts = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
+        backup_path = f"{self.raw_base_path}/backup/{ts}/{self.source_system}/{self.source_class}"
+        print(f"Writing {input_table} to {backup_path}")
+        bronze_df = self.spark.sql(f"select * from {input_table}")
+        (
+            bronze_df
+            .drop("_rescued_data")
+            .drop("_ingest_time")
+            .drop("_quarantined")
+            .write
+            .mode("overwrite")
+            .parquet(backup_path)
+        )
+        return backup_path
+
+    def ingest_historical(
+        self,
+        data_type = None,
+        environment = None,
+        input_location = None,
+        input_format = "parquet"
+        ):
+        """ Ingest historical data """
+        if data_type:
+            warnings.warn("Parameter data_type is deprecated",
+                category=DeprecationWarning,
+                stacklevel=2)
+        if environment:
+            warnings.warn("Parameter environment is deprecated",
+                category=DeprecationWarning,
+                stacklevel=2)
+        if not input_location:
+            raise Exception("input_location must be psecified")
+
+        target_location = f"{self.raw_base_path}/{self.source_system}/{self.source_class}_historical"
+        print(f"Adding {input_location} to {target_location}")
+        input_df = self.spark.read.format(input_format).load(input_location)
+        (
+            input_df
+            .write
+            .mode("append")
+            .parquet(target_location)
+        )
